# fastapi_backend/app/routes/conversion.py
from uuid import UUID
from pydantic import BaseModel

from pathlib import Path
from fastapi import APIRouter, Depends, HTTPException, Query, UploadFile, File
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy.future import select
from sqlalchemy import Column, String, Integer, ForeignKey
from sqlalchemy.orm import DeclarativeBase, relationship

from app.models import DocPipelines
from app.database import User, get_async_session
from app.users import current_active_user

from typing import List, Dict, Any
from uuid import uuid4
import json
import os
import pandas as pd

# helper for removing "color" from stored pipelines
from app.rag_services.helpers import load_pipeline, ExtractionError

# Indexing service
from app.rag_services.indexing_service import run_conversion, run_md_processing

# Markdown generator
from app.generate_markdown import generate_markdown_from_log, find_session_id



router = APIRouter(tags=["conversion"])






MethodSpec = Dict[str, Any]



@router.get("/{doc_id}/data", response_model=MethodSpec)
async def read_conversion_pipeline(
    doc_id: UUID,
    db: AsyncSession = Depends(get_async_session),
    user: User = Depends(current_active_user),
):
    row = await DocPipelines.get_row(where_dict={"user_id": user.id, "doc_id": doc_id}, db=db)


    if row.conversion_pipeline is None:
        # Return default empty pipeline if none exists
        return {}

    conversion_pipeline = json.loads(row.conversion_pipeline)

    return conversion_pipeline




@router.post("/{doc_id}/data")
async def add_conversion_pipeline(
    doc_id: UUID,
    pipeline: MethodSpec,
    db: AsyncSession = Depends(get_async_session),
    user: User = Depends(current_active_user),
):
    # First we delete current pipeline if it's set
    row = await DocPipelines.get_row(where_dict={"user_id": user.id, "doc_id": doc_id}, db=db)
    row.conversion_pipeline = json.dumps(pipeline)

    await db.commit()
    await db.refresh(row)

    return {
        "status": "ok"
    }



@router.post("/{doc_id}/run")
async def run_conversion_pipeline(
    doc_id: UUID,
    db: AsyncSession = Depends(get_async_session),
    user: User = Depends(current_active_user),
):

    try:

        # 1. filter: this endpoint is only triggered when a pipeline is fetched or created from the UI
        row = await DocPipelines.get_row(where_dict={"user_id": user.id, "doc_id": doc_id}, db=db)


        # 2. filter: output error if the pipeline is created but not saved, and there is no previous pipeline
        if not row.conversion_pipeline:
            raise HTTPException(status_code=404, detail="No pipeline was saved")

        # Run conversion
        conversion_pipeline = json.loads(row.conversion_pipeline)

        await run_conversion(conversion_pipeline, user.id, doc_id, db)

        # After conversion

        # Set converted=1
        row.converted = True

        row.chunked = False

        # NEXT: Set exported=0
        row.exported = False

        await db.commit()


        return {
            "status": "ok"
        }

    except ExtractionError as e:
        raise HTTPException(status_code=e.status_code, detail=e.message)




# ---------- ALL DOCs ----------
@router.post("/run")
async def convert_all(
    db: AsyncSession = Depends(get_async_session),
    user: User = Depends(current_active_user),
):
    # Avoid converted=1 to be converted again.

    rows, _ = await DocPipelines.get_all(where_dict={"user_id": user.id, "converted": 0}, db=db)

    doc_ids = [row["doc_id"] for row in rows]

    for doc_id in doc_ids:
        row = await DocPipelines.get_row(where_dict={"user_id": user.id, "doc_id": doc_id}, db=db)

        if row.conversion_pipeline:
            conversion_pipeline = json.loads(row.conversion_pipeline)

            await run_conversion(conversion_pipeline, user.id, doc_id, db)

            # After conversion

            # Set converted=1
            row.converted = True

            # Set chunked=0
            row.chunked = False

            # Set exported=0
            row.exported = False

    await db.commit()


    return {
        "status": "ok"
    }


